+++ R/PivotThemes.R
@@ -37,30 +37,33 @@ getDefaultTheme <- function(parentPivot, themeName="default") {
       "border-collapse"="collapse"
     ))
   pivotStyles$addStyle(styleName="ColumnHeader", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.75em",
       padding="2px",
       border="1px solid lightgray",
       "vertical-align"="middle",
       "text-align"="center",
       "font-weight"="bold",
-      "background-color"="#F2F2F2"
+      "background-color"="#F2F2F2",
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="RowHeader", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.75em",
       padding="2px 8px 2px 2px",
       border="1px solid lightgray",
       "vertical-align"="middle",
       "text-align"="left",
       "font-weight"="bold",
-      "background-color"="#F2F2F2"
+      "background-color"="#F2F2F2",
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="Cell", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.75em",
       padding="2px 2px 2px 8px",
       border="1px solid lightgray",
+      "vertical-align"="middle",
       "text-align"="right"
     ))
   pivotStyles$tableStyle <- "Table"
@@ -90,31 +93,34 @@ getLargePlainTheme <- function(parentPivot, themeName="largeplain") {
       "border-collapse"="collapse"
     ))
   pivotStyles$addStyle(styleName="ColumnHeader", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.875em",
       padding="4px",
       "min-width"="100px",
       border="1px solid lightgray",
       "vertical-align"="middle",
       "text-align"="center",
-      "font-weight"="bold"
+      "font-weight"="bold",
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="RowHeader", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.875em",
       padding="4px",
       "min-width"="100px",
       border="1px solid lightgray",
       "vertical-align"="middle",
       "text-align"="left",
-      "font-weight"="bold"
+      "font-weight"="bold",
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="Cell", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.875em",
       padding="4px",
       "min-width"="100px",
       border="1px solid lightgray",
+      "vertical-align"="middle",
       "text-align"="right"
     ))
   pivotStyles$tableStyle <- "Table"
@@ -144,29 +150,33 @@ getCompactTheme <- function(parentPivot, themeName="compact") {
       "border-collapse"="collapse"
     ))
   pivotStyles$addStyle(styleName="ColumnHeader", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.625em",
       padding="2px",
       border="1px solid lightgray",
       "vertical-align"="middle",
       "text-align"="center",
       "font-weight"="bold",
-      "background-color"="#F2F2F2"
+      "background-color"="#F2F2F2",
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="RowHeader", list(
-      "font-family"="arial",
+      "font-family"="Arial",
       "font-size"="0.625em",
       padding="2px 4px 2px 2px",
       border="1px solid lightgray",
       "vertical-align"="middle",
       "text-align"="left",
       "font-weight"="bold",
-      "background-color"="#F2F2F2"
+      "background-color"="#F2F2F2",
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="Cell", list(
-      font="0.625em arial",
+      "font-family"="Arial",
+      "font-size"="0.625em",
       padding="2px 2px 2px 6px",
       border="1px solid lightgray",
+      "vertical-align"="middle",
       "text-align"="right"
     ))
   pivotStyles$tableStyle <- "Table"
@@ -211,7 +221,8 @@ getSimpleColoredTheme <- function(parentPivot, themeName="coloredTheme", colors,
       "text-align"="center",
       "font-weight"="bold",
       color=colors$headerColor,
-      "background-color"=colors$headerBackgroundColor
+      "background-color"=colors$headerBackgroundColor,
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="RowHeader", list(
       "font-family"=fontName,
@@ -222,13 +233,15 @@ getSimpleColoredTheme <- function(parentPivot, themeName="coloredTheme", colors,
       "text-align"="left",
       "font-weight"="bold",
       color=colors$headerColor,
-      "background-color"=colors$headerBackgroundColor
+      "background-color"=colors$headerBackgroundColor,
+      "xl-wrap-text"="wrap"
     ))
   pivotStyles$addStyle(styleName="Cell", list(
       "font-family"=fontName,
       "font-size"="0.75em",
       padding="2px 2px 2px 8px",
       "border"=paste0("1px solid ", colors$borderColor),
+      "vertical-align"="middle",
       "text-align"="right",
       color=colors$cellColor,
       "background-color"=colors$cellBackgroundColor
@@ -238,6 +251,7 @@ getSimpleColoredTheme <- function(parentPivot, themeName="coloredTheme", colors,
       "font-size"="0.75em",
       padding="2px 2px 2px 8px",
       "border"=paste0("1px solid ", colors$borderColor),
+      "vertical-align"="middle",
       "text-align"="right",
       color=colors$totalColor,
       "background-color"=colors$totalBackgroundColor